# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test task API Documentation",
        "description": "Test task API Documentation Description",
        "contact": {
            "name": "Swagger API Team"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/",
            "description": "Local API server"
        }
    ],
    "paths": {
...
        "/api/createImage": {
            "post": {
                "tags": [
                    "Image Route"
                ],
                "summary": "Image add",
                "description": "Image post.",
                "operationId": "ImageCreate",
                "parameters": [
                    {
                        "name": "image",
                        "in": "query",
                        "description": "Image name",
                        "required": true,
                        "schema": {
                            "type": "file"
                        },
                        "example": "Image Name"
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Brand or Branch",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Brand & Branch Name"
                    },
                    {
                        "name": "nameOfBrand",
                        "in": "query",
                        "description": "Brand",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Brand Name"
                    },
                    {
                        "name": "nameOfBranch",
                        "in": "query",
                        "description": "Branch",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Branch Name"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized!"
                    },
                    "422": {
                        "description": "Validation error!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}